Drop suppress_errors where possible and minor refactors
- drop suppress_errors in functions that are well-suited for, or
  even already are, returning a None
- handle both context.json and context for the `is_profile_exists` check
- use try/except IOError instead of isfile for the get_active_profile
  (EAFP is better than LBYL wrt. possible race conditions)
- drop a few unittests that were specifically testing for suppress_errors
- drop the specialcased local profile in get_profile_context (and a test)
  because instead, we just don't call it for local profiles instead.
- env.delete_profile silently Just Deletes™ and does nothing if it's
  already deleted. This is because env.delete_profile is also used
  internally (eg. `cfy profiles set --profile-name`). For the case
  where we _do_ want to throw an error (`cfy prof delete` on a
  nonexistent profile), throw the error from inside the cli command
  itself. (also made it correctly return nonzero in that case,
  and updated a unittest to reflect that)

+++ cloudify_cli/commands/profiles.py
@@ -298,11 +298,11 @@ def delete(profile_name, logger):
     `PROFILE_NAME` is the IP of the manager the profile manages.
     """
     logger.info('Deleting profile {0}...'.format(profile_name))
-    try:
-        env.delete_profile(profile_name)
-        logger.info('Profile deleted')
-    except CloudifyCliError as ex:
-        logger.info(str(ex))
+    if not env.is_profile_exists(profile_name):
+        raise CloudifyCliError('Profile {0} does not exist'
+                               .format(profile_name))
+    env.delete_profile(profile_name)
+    logger.info('Profile deleted')
 
 
 def set_profile(profile_name,
